+++ app/concerns/koudoku/subscription.rb
@@ -6,7 +6,7 @@ module Koudoku::Subscription
     # We don't store these one-time use tokens, but this is what Stripe provides
     # client-side after storing the credit card information.
     attr_accessor :credit_card_token
-    
+
     belongs_to :plan
 
     # update details.
@@ -14,7 +14,7 @@ module Koudoku::Subscription
     def processing!
 
       # if their package level has changed ..
-      if changing_plans? 
+      if changing_plans?
 
         prepare_for_plan_change
 
@@ -34,7 +34,7 @@ module Koudoku::Subscription
             prepare_for_upgrade if upgrading?
 
             # update the package level with stripe.
-            customer.update_subscription(:plan => self.plan.stripe_id)
+            customer.update_subscription(:plan => self.plan.stripe_id, :prorate => Koudoku.prorate)
 
             finalize_downgrade! if downgrading?
             finalize_upgrade! if upgrading?
@@ -71,7 +71,8 @@ module Koudoku::Subscription
               customer_attributes = {
                 description: subscription_owner_description,
                 email: subscription_owner_email,
-                plan: plan.stripe_id
+                plan: plan.stripe_id,
+                card: credit_card_token # obtained with Stripe.js
               }
               if plan.price > 0.0 and credit_card_token.present?
                 customer_attributes[:card] = credit_card_token # obtained with Stripe.js
@@ -83,10 +84,15 @@ module Koudoku::Subscription
                   customer_attributes[:trial_end] = coupon.free_trial_ends.to_i
                 end
               end
+              
+              customer_attributes[:coupon] = @coupon_code if @coupon_code 
 
               # create a customer at that package level.
               customer = Stripe::Customer.create(customer_attributes)
-              
+
+              finalize_new_customer!(customer.id, plan.price)
+              customer.update_subscription(:plan => self.plan.stripe_id, :prorate => Koudoku.prorate)
+
             rescue Stripe::CardError => card_error
               errors[:base] << card_error.message
               card_was_declined
@@ -114,12 +120,12 @@ module Koudoku::Subscription
         end
 
         finalize_plan_change!
-        
+
       # if they're updating their credit card details.
       elsif self.credit_card_token.present?
-        
+
         prepare_for_card_update
-        
+
         # fetch the customer.
         customer = Stripe::Customer.retrieve(self.stripe_id)
         customer.card = self.credit_card_token
@@ -134,10 +140,8 @@ module Koudoku::Subscription
     end
 
   end
-
-  module ClassMethods
-  end
-
+  
+  
   def describe_difference(plan_to_describe)
     if plan.nil?
       if persisted?
@@ -145,7 +149,7 @@ module Koudoku::Subscription
       else
         if Koudoku.free_trial?
           "Start Trial"
-        else 
+        else
           "Upgrade"
         end
       end
@@ -157,6 +161,12 @@ module Koudoku::Subscription
       end
     end
   end
+  
+  # Set a Stripe coupon code that will be used when a new Stripe customer (a.k.a. Koudoku subscription)
+  # is created
+  def coupon_code=(new_code)
+    @coupon_code = new_code
+  end
 
   # Pretty sure this wouldn't conflict with anything someone would put in their model
   def subscription_owner
@@ -165,14 +175,19 @@ module Koudoku::Subscription
     send Koudoku.subscriptions_owned_by
   end
 
+  def subscription_owner=(owner)
+    # e.g. @subscription.user = @owner
+    send Koudoku.owner_assignment_sym, owner
+  end
+
   def subscription_owner_description
     # assuming owner responds to name.
     # we should check for whether it responds to this or not.
-    "#{subscription_owner.id}"
+    "#{subscription_owner.try(:name) || subscription_owner.try(:id)}"
   end
 
   def subscription_owner_email
-    nil
+    "#{subscription_owner.try(:email)}"
   end
 
   def changing_plans?
@@ -202,7 +217,7 @@ module Koudoku::Subscription
 
   def prepare_for_cancelation
   end
-  
+
   def prepare_for_card_update
   end
 
@@ -210,6 +225,9 @@ module Koudoku::Subscription
   end
 
   def finalize_new_subscription!
+  end
+
+  def finalize_new_customer!(customer_id, amount)
   end
 
   def finalize_upgrade!
@@ -226,14 +244,14 @@ module Koudoku::Subscription
 
   def card_was_declined
   end
-  
+
   # stripe web-hook callbacks.
   def payment_succeeded(amount)
   end
-  
+
   def charge_failed
   end
-  
+
   def charge_disputed
   end