+++ lib/generators/koudoku/templates/app/models/coupon.rb
@@ -1,0 +1,5 @@
+class Coupon < ActiveRecord::Base
+  
+  has_many :subscriptions
+
+end