+++ lib/koudoku.rb
@@ -1,12 +1,9 @@
 require "koudoku/engine"
 require "generators/koudoku/install_generator"
 require "generators/koudoku/views_generator"
+require 'stripe_event'
 
 module Koudoku
-
-  mattr_accessor :webhooks_api_key
-  @@webhooks_api_key = nil
-  
   mattr_accessor :subscriptions_owned_by
   @@subscriptions_owned_by = nil
   
@@ -18,6 +15,13 @@ module Koudoku
   
   mattr_accessor :free_trial_length
   @@free_trial_length = nil
+
+  mattr_accessor :prorate
+  @@prorate = true
+
+  def self.webhooks_api_key=(key)
+    raise "Koudoku no longer uses an API key to secure webhooks, please delete the line from \"config/initializers/koudoku.rb\""
+  end
   
   def self.setup
     yield self
@@ -33,18 +37,16 @@ module Koudoku
   
   # e.g. :user_id
   def self.owner_id_sym
-    # e.g. :user_id
-    (Koudoku.subscriptions_owned_by.to_s + '_id').to_sym
+    :"#{Koudoku.subscriptions_owned_by}_id"
   end
   
+  # e.g. :user=
   def self.owner_assignment_sym
-    # e.g. :user=
-    (Koudoku.subscriptions_owned_by.to_s + '=').to_sym
+    :"#{Koudoku.subscriptions_owned_by}="
   end
 
-  # e.g. Users
+  # e.g. User
   def self.owner_class
-    # e.g. User
     Koudoku.subscriptions_owned_by.to_s.classify.constantize
   end
   
@@ -52,4 +54,20 @@ module Koudoku
     free_trial_length.to_i > 0
   end
   
+  
+  #
+  # STRIPE_EVENT section
+  #
+  def self.subscribe(name, callable = Proc.new)
+    StripeEvent.subscribe(name, callable)
+  end
+
+  def self.instrument(name, object)
+    StripeEvent.backend.instrument(StripeEvent.namespace.call(name), object)
+  end
+
+  def self.all(callable = Proc.new)
+    StripeEvent.all(callable)
+  end
+
 end